serialize_test/requirements.txt
lxml
jsonpickle
protobuf==3.17.1
grpcio-tools==1.44.0
avro
pyyaml
msgpack
